CXC2026_Django_Backend/requirements.txt
django>=5.1,<5.2
gunicorn>=23,<24
dj-database-url>=2,<3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
